\name{custom.featureTable}
\alias{custom.featureTable}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Import Custom Feature Table
}
\description{
Perform IPA analysis using user provided feature table in CSV format generated by MS-DIAL, MZmine2, etc instead of using XCMS feature table. Also annotates each feature as either level 1 or 2.
}
\usage{
custom.featureTable(dir, FTdir, mz.tol = 10)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dir}{
Directory of folder containing all .mzXML files of interest. This folder must contain only relevant .mzXML files of DDA sample(s).
}
  \item{FTdir}{
Directory containing feature table in CSV format.
}
  \item{mz.tol}{
Mass tolerance in ppm used to identify level 1 features by matching the feature's mz against all ms2 precursor mz. Default = 10.
}
}
\details{
The number of mzXML files to process must be equal to the number of input CSV files where each CSV file contains features extracted from each mzXML file. Note: for multi-sample analysis, the feature table provided must not be aligned. The existing feature table must be in CSV format and contains only the following columns in order: "mz", "rt", "rtmin", "rtmax", "intensity".
}
\value{
Returns a feature table in dataframe format, containing the following columns in order: "mz", "rt", "rtmin", "rtmax", "maxo", "sample", "level" where "maxo" is the feature peak intensity. Also returns an XCMS object that is used by functions later in the IPA workflow.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Sam Shen, Jian Guo, Tao Huan
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
library(IPA)
dir = "X:/Users/Sam_Shen/IPAtest_20210330/singleDDA"
FTdir = "X:/Users/Sam_Shen/IPAtest_20210330/singleDDA"
featureTable <- read.featureTable(dir = dir, FTdir = FTdir)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
